Make it its own class
